+++ test/IntegrationTests/Helpers/DockerFileHelper.cs
@@ -7,9 +7,9 @@ internal static class DockerFileHelper
 {
     public static string ReadImageFrom(string dockerfile)
     {
-        string raw = File.ReadAllText($"./docker/{dockerfile}");
+        var raw = File.ReadAllText($"./docker/{dockerfile}");
 
-        // Exlcude FROM
+        // Exclude FROM
         return raw.Substring(4).Trim();
     }
 }